+++ aws_lambda_powertools/event_handler/openapi/params.py
@@ -6,12 +6,15 @@ from typing import Annotated, Any, Callable, Dict, List, Optional, Tuple, Union,
 from pydantic import BaseConfig
 from pydantic.fields import FieldInfo, ModelField, Required, Undefined
 from pydantic.schema import get_annotation_from_field_info
+from pydantic.version import VERSION as PYDANTIC_VERSION
 
-from aws_lambda_powertools.event_handler.openapi import Example
+from aws_lambda_powertools.event_handler.openapi.utils import CacheKey
 
 """
 This turns the low-level function signature into typed, validated Pydantic models for consumption.
 """
+
+PYDANTIC_V2 = PYDANTIC_VERSION.startswith("2.")
 
 
 class Dependant:
@@ -55,7 +58,7 @@ class Dependant:
         # Store the path to be able to re-generate a dependable from it in overrides
         self.path = path
         # Save the cache key at creation to optimize performance
-        self.cache_key = self.call
+        self.cache_key: CacheKey = self.call
 
 
 class ParamTypes(Enum):
@@ -65,6 +68,10 @@ class ParamTypes(Enum):
     cookie = "cookie"
 
 
+# MAINTENANCE: update when deprecating Pydantic v1, remove this alias
+_Unset: Any = Undefined
+
+
 class Param(FieldInfo):
     in_: ParamTypes
 
@@ -72,12 +79,12 @@ class Param(FieldInfo):
         self,
         default: Any = Undefined,
         *,
-        default_factory: Union[Callable[[], Any], None] = Undefined,
+        default_factory: Union[Callable[[], Any], None] = _Unset,
         annotation: Optional[Any] = None,
         alias: Optional[str] = None,
-        alias_priority: Union[int, None] = Undefined,
-        # TODO: update when deprecating Pydantic v1, import these types
-        # validation_alias: str | AliasPath | AliasChoices | None
+        alias_priority: Union[int, None] = _Unset,
+        # MAINTENANCE: update when deprecating Pydantic v1, import these types
+        # MAINTENANCE: validation_alias: str | AliasPath | AliasChoices | None
         validation_alias: Union[str, None] = None,
         serialization_alias: Union[str, None] = None,
         title: Optional[str] = None,
@@ -90,13 +97,12 @@ class Param(FieldInfo):
         max_length: Optional[int] = None,
         pattern: Optional[str] = None,
         discriminator: Union[str, None] = None,
-        strict: Union[bool, None] = Undefined,
-        multiple_of: Union[float, None] = Undefined,
-        allow_inf_nan: Union[bool, None] = Undefined,
-        max_digits: Union[int, None] = Undefined,
-        decimal_places: Union[int, None] = Undefined,
+        strict: Union[bool, None] = _Unset,
+        multiple_of: Union[float, None] = _Unset,
+        allow_inf_nan: Union[bool, None] = _Unset,
+        max_digits: Union[int, None] = _Unset,
+        decimal_places: Union[int, None] = _Unset,
         examples: Optional[List[Any]] = None,
-        openapi_examples: Optional[Dict[str, Example]] = None,
         deprecated: Optional[bool] = None,
         include_in_schema: bool = True,
         json_schema_extra: Union[Dict[str, Any], None] = None,
@@ -104,7 +110,7 @@ class Param(FieldInfo):
     ):
         self.deprecated = deprecated
         self.include_in_schema = include_in_schema
-        self.openapi_examples = openapi_examples
+
         kwargs = dict(
             default=default,
             default_factory=default_factory,
@@ -128,9 +134,23 @@ class Param(FieldInfo):
             kwargs["examples"] = examples
 
         current_json_schema_extra = json_schema_extra or extra
-        kwargs["regex"] = pattern
-        kwargs.update(**current_json_schema_extra)
-        use_kwargs = {k: v for k, v in kwargs.items() if v is not Undefined}
+        if PYDANTIC_V2:
+            kwargs.update(
+                {
+                    "annotation": annotation,
+                    "alias_priority": alias_priority,
+                    "validation_alias": validation_alias,
+                    "serialization_alias": serialization_alias,
+                    "strict": strict,
+                    "json_schema_extra": current_json_schema_extra,
+                },
+            )
+            kwargs["pattern"] = pattern
+        else:
+            kwargs["regex"] = pattern
+            kwargs.update(**current_json_schema_extra)
+
+        use_kwargs = {k: v for k, v in kwargs.items() if v is not _Unset}
 
         super().__init__(**use_kwargs)
 
@@ -145,12 +165,12 @@ class Path(Param):
         self,
         default: Any = ...,
         *,
-        default_factory: Union[Callable[[], Any], None] = Undefined,
+        default_factory: Union[Callable[[], Any], None] = _Unset,
         annotation: Optional[Any] = None,
         alias: Optional[str] = None,
-        alias_priority: Union[int, None] = Undefined,
-        # TODO: update when deprecating Pydantic v1, import these types
-        # validation_alias: str | AliasPath | AliasChoices | None
+        alias_priority: Union[int, None] = _Unset,
+        # MAINTENANCE: update when deprecating Pydantic v1, import these types
+        # MAINTENANCE: validation_alias: str | AliasPath | AliasChoices | None
         validation_alias: Union[str, None] = None,
         serialization_alias: Union[str, None] = None,
         title: Optional[str] = None,
@@ -163,13 +183,12 @@ class Path(Param):
         max_length: Optional[int] = None,
         pattern: Optional[str] = None,
         discriminator: Union[str, None] = None,
-        strict: Union[bool, None] = Undefined,
-        multiple_of: Union[float, None] = Undefined,
-        allow_inf_nan: Union[bool, None] = Undefined,
-        max_digits: Union[int, None] = Undefined,
-        decimal_places: Union[int, None] = Undefined,
+        strict: Union[bool, None] = _Unset,
+        multiple_of: Union[float, None] = _Unset,
+        allow_inf_nan: Union[bool, None] = _Unset,
+        max_digits: Union[int, None] = _Unset,
+        decimal_places: Union[int, None] = _Unset,
         examples: Optional[List[Any]] = None,
-        openapi_examples: Optional[Dict[str, Example]] = None,
         deprecated: Optional[bool] = None,
         include_in_schema: bool = True,
         json_schema_extra: Union[Dict[str, Any], None] = None,
@@ -202,7 +221,6 @@ class Path(Param):
             decimal_places=decimal_places,
             deprecated=deprecated,
             examples=examples,
-            openapi_examples=openapi_examples,
             include_in_schema=include_in_schema,
             json_schema_extra=json_schema_extra,
             **extra,
@@ -214,12 +232,12 @@ class Query(Param):
 
     def __init__(
         self,
-        default: Any = Undefined,
+        default: Any = _Unset,
         *,
-        default_factory: Union[Callable[[], Any], None] = Undefined,
+        default_factory: Union[Callable[[], Any], None] = _Unset,
         annotation: Optional[Any] = None,
         alias: Optional[str] = None,
-        alias_priority: Union[int, None] = Undefined,
+        alias_priority: Union[int, None] = _Unset,
         validation_alias: Union[str, None] = None,
         serialization_alias: Union[str, None] = None,
         title: Optional[str] = None,
@@ -232,13 +250,12 @@ class Query(Param):
         max_length: Optional[int] = None,
         pattern: Optional[str] = None,
         discriminator: Union[str, None] = None,
-        strict: Union[bool, None] = Undefined,
-        multiple_of: Union[float, None] = Undefined,
-        allow_inf_nan: Union[bool, None] = Undefined,
-        max_digits: Union[int, None] = Undefined,
-        decimal_places: Union[int, None] = Undefined,
+        strict: Union[bool, None] = _Unset,
+        multiple_of: Union[float, None] = _Unset,
+        allow_inf_nan: Union[bool, None] = _Unset,
+        max_digits: Union[int, None] = _Unset,
+        decimal_places: Union[int, None] = _Unset,
         examples: Optional[List[Any]] = None,
-        openapi_examples: Optional[Dict[str, Example]] = None,
         deprecated: Optional[bool] = None,
         include_in_schema: bool = True,
         json_schema_extra: Union[Dict[str, Any], None] = None,
@@ -269,7 +286,6 @@ class Query(Param):
             decimal_places=decimal_places,
             deprecated=deprecated,
             examples=examples,
-            openapi_examples=openapi_examples,
             include_in_schema=include_in_schema,
             json_schema_extra=json_schema_extra,
             **extra,
